Forms Hooked up, minor styling
