+++ Source/nt.Client/Features/WebThree/Components/NftTemplates/BillOfLading/BillOfLadingTemplate.razor.cs
@@ -4,7 +4,6 @@
   using nt.Client.Features.Base.Components;
   using nt.Client.Features.WebThree.Actions.AddFormDataToState;
   using System;
-  using System.Threading.Tasks;
 
   public class BillOfLadingTemplateModel : BaseComponent
   {
