+++ Source/nt.Client/Features/WebThree/Components/NftTemplates/DocumentComponent.razor
@@ -1,9 +1,28 @@
-﻿@inherits  DocumentComponentModel
+﻿@inherits DocumentComponentModel
 
 <div>
-  @if(NftId == 5)
+  @switch (NftId)
   {
-    <nt.Client.Features.WebThree.Components.NftTemplates.PurchaseOrder.PurchaseOrderTemplate />
-  }
+
+    case 5:
+      <nt.Client.Features.WebThree.Components.NftTemplates.PurchaseRequest.PurchaseRequestTemplate />;
+      break;
+
+    case 4:
+      <nt.Client.Features.WebThree.Components.NftTemplates.BillOfLading.BillOfLadingTemplate />;
+      break;
+
+    case 3:
+      <nt.Client.Features.WebThree.Components.NftTemplates.RequestForProposal.RequestForProposalTemplate />;
+      break;
+
+    case 2:
+      <nt.Client.Features.WebThree.Components.NftTemplates.RequestForQuotation.RequestForQuotationTemplate/>;
+      break;
+
+    default:
+  <nt.Client.Features.WebThree.Components.NftTemplates.PurchaseOrder.PurchaseOrderTemplate />;
+      break;
+  } 
 </div>
