--- Source/nt.Client/Features/WebThree/Components/NftTemplates/FormDto.cs
@@ -1,10 +1,0 @@
-﻿namespace nt.Client.Features.WebThree.Components.NftTemplates
-{
-  public class FormDto
-  {
-    public ImmutableObjectBase FormValues { get; set; }
-
-    public string MutableDataString { get; set; }
-
-  }
-}
