+++ Source/nt.Client/Features/WebThree/Components/NftTypeListComponent.razor
@@ -12,6 +12,9 @@
   </div>
   <div class="d-flex justify-content-center w-50 mx-auto">
     <div class="col m-1 text-center bg-light">
+      <p class="font-weight-bold">Nft Id:</p> @CurrentNftTemplate.NftId
+    </div>
+    <div class="col m-1 text-center bg-light">
       <p class="font-weight-bold">Name:</p> @CurrentNftTemplate.Name
     </div>
     <div class="col m-1 text-center bg-light">
